Make all BlockStore iterations use CloseableIterator
### What changes are proposed in this pull request?

For all callers of BlockStore iterators, force them to get a
CloseableIterator and use try-with-resource pattern.

### Why are the changes needed?

Even after https://github.com/Alluxio/alluxio/pull/14964/files, there
are still operation that directly iterate the BlockStore without using
the CloseableIterator and try-with-resource pattern. This change is to
address newly found ones.

The root cause is `BlockStore extends Iterable` which gives `Iterator`
to callers. Callers are not aware that the iterators are essentially
`CloseableIterator` and do not close them. It's not easy to close
either, the caller must first do a checked cast to `CloseableIterator`,
and the whole process just does not support try-with-resource pattern.

So this change totally removed interface of `Iterator<Block> iterator()`
and provides a replacement `CloseableIterator<Block>
getCloseableIterator()`, so the callers can directly use
try-with-resource pattern.

### Does this PR introduce any user facing changes?

NA

pr-link: Alluxio/alluxio#15067
change-id: cid-e44ffa2ac9e27cb0c20b7c5b1f430699f1a0d589