Add a metric to show the count of blocks to remove
### What changes are proposed in this pull request?

Add a metric to obtain the number of data blocks that have not been
cleaned up temporarily.

### Why are the changes needed?

When the `free` command is executed, since the behavior of free to clear
data blocks is asynchronous, even if free returns successfully, there
are still data blocks that have not been cleared temporarily. If the
`load` operation is executed again at this time, it will be found that
the load execution speed is very fast, which is inaccurate. By
increasing this metric, we can monitor the completion progress of the
free command to a certain extent. Since the collection of this metric
may affect performance, I set a switch to turn it on only when needed.

### Does this PR introduce any user facing changes?
No

pr-link: Alluxio/alluxio#15023
change-id: cid-6d72f87c41e1a9f8c074ce2529d3727e8cf7ec77