Add a configuration to set ufs fetch timeout
### What changes are proposed in this pull request?

This change makes the UFS fetch operation timeout configurable.

Also a minor change which uses the sized constructor for HashMap.

### Why are the changes needed?

Some UFSes can be bad in worst response time so the fetch operation can
timeout easily when the timeout is just 100ms. This might be most useful
in the development stage, where the user is accessing a remote UFS like
S3 from a random location, so the response time can be randomly bad.

### Does this PR introduce any user facing changes?

This adds a new property key for the user to configure if really needs
to.

pr-link: Alluxio/alluxio#15065
change-id: cid-a46a9eb20327c089576aacf2c55e794a16063c4b