Replace PowerMock with Mockito in LogLevelTest
### What changes are proposed in this pull request?

Replace PowerMock with Mockito in LogLevelTest.

### Why are the changes needed?

Part of #15003.

This PR intentionally limits the scope of change to a single low-impact
test, so that we can first add needed dependencies and see if it works,
then proceed to replace more PowerMock in other tests.

### Does this PR introduce any user facing changes?

No.

### Why not set mockito-inline in all child-projects but only apply this
dependency to the `shell` project?

`powermock` is still used in other child projects. If the
`mockito-inline` dependency is enabled for all child projects in root
pom, because of https://github.com/powermock/powermock/issues/992, there
will be conflicts between powermock and mockito-inline.

The plan is to progressively replace powermock module-by-module, then
remove powermock from root pom and enable mockito-inline there.

pr-link: Alluxio/alluxio#15091
change-id: cid-e2b32959d544ec2697d498a37e0c53186b18267e