Ignore errors while closing the UFS journal log writer's internal stream
When master process is paused (via sigkill) for a long time, the
internal hdfs-client can get confused about the data-node states and
throw runtime exception. In an FT cluster, this currently means if a
master encounters this during losePrimacy time, it'll never come back
up.

Overall, it's OK to ignore internal streaming errors during `close()`.
Because the next time it'll either be a read stream or a new write
stream.

pr-link: Alluxio/alluxio#15132
change-id: cid-7d75a917eeaa47f1d426606c454b8a0918ab38f9