Add dir-start-index fuseStressBench
When generating data in fuse stressbench, user can choose the directory
names to start from a certain index.

With this change, user is able to generate all test data in different
batches. For example, for the first batch we set `--num-dirs 10` and
`--dir-start-index 0`, and for the second batch we set `--num-dirs 10`
and `--dir-start-index 10`, we are able to generate 20-dirs test files
that is recognizable by reading stress bench.

pr-link: Alluxio/alluxio#15135
change-id: cid-46b6a2391fda7305ee267f917d9d7456d612f173