[DOCFIX] Update k8s doc
A minor update to add `--` in the `kubectl` command in the doc.

We can still run the command as expected if we don't put the `--` before
the command, but we would get a warning `kubectl exec [POD] [COMMAND] is
DEPRECATED and will be removed in a future version. Use kubectl exec
[POD] -- [COMMAND] instead. `

pr-link: Alluxio/alluxio#15112
change-id: cid-d319bf30a52b4bb07109a539381a31ca9c93f08e