Fix java installation directory in Dockerfile-dev
there's a line in the dockerfiles to disable jvm dns cache by appending
to this file within the java installation directory
`/usr/lib/jvm/java-1.8-openjdk/jre/lib/security/java.security`. sometime
between the morning of 3/23 and 3/24, something changed such that this
path doesn't exist. maybe due to an upgrade in either the java version
or the way the OS creates the symlinks, but now the path appends `.0` to
the end of `java-1.8` ->
`/usr/lib/jvm/java-1.8.0-openjdk/jre/lib/security/java.security`.

note this only affects java 8 on the centos base image; java 11 on the
same image and java 8 on the alpine image are not affected

pr-link: Alluxio/alluxio#15199
change-id: cid-515ab6f812deec092d41cfbdc6a6d908d867b3c1