Reduce copy in UfsStatusCache
### What changes are proposed in this pull request?

This change avoids a few copies in UfsStatusCache, by changing the
method interfaces.

### Why are the changes needed?

The copies will consume a lot of extra memory, especially when we sync
large directories

For a `UfsStatus[]` of length `n`, the memory save is one
`Arrays.asList()` + one `HashSet`.

The cost of `Arrays.asList()` is documented as just creating a wrapper
in oracle doc

https://docs.oracle.com/javase/8/docs/api/java/util/Arrays.html#asList-T...-
But if you go into the source when you are using adoptopenjdk8, it's
actually creating an `ArrayList<>(arr)`. If the real cost is an
ArrayList, then the cost is around `8 bytes *n`.

The cost of `HashSet` is approximately also 8 bytes per element but the
number of elements is the ceiling power of 2. So it's between `[8 bytes
* n, 16 bytes * n)`.

So the two combined will be `[16B * n, 24B * n)`. But these are just
temporary data structures and likely just GC-ed during young GC.

Don't forget the CPU overhead of collection creation, obj creation etc.
Although I don't have a micro benchmark, the benefit there is probably
more significant.

pr-link: Alluxio/alluxio#15160
change-id: cid-0907fccc7352728f134ebf7582ac72216aaa2f7a