Avoid ArrayList resizing in getFileBlockInfoListInternal
### What changes are proposed in this pull request?

Init the ArrayList with size to avoid internal resizing.

If no size is given on ArrayList construction, internally I believe
ArrayList init with 0 size and on adding the 1st element, it allocates a
10 element array. So when the average block count in Alluxio is
significantly less than 10, we can save some space on ArrayList init.
And when the average block count in Alluxio is significantly over 10, we
half the overall mem allocation of ArrayList by avoiding resizing.

However, the save is mostly mem allocations in the Young Gen and Young
GC, so it's not going to be that observable.

### Does this PR introduce any user facing changes?

NA

pr-link: Alluxio/alluxio#15213
change-id: cid-35f213fbc5f934ce01af94099e3d6ed282c1c476