Save one list copy in rm
### What changes are proposed in this pull request?

1. Refactor the rm to avoid one list copy
2. Better init sizing to avoid some copy from ArrayList auto-sizing
3. Setting unused positions to `null` will help GC

### Why are the changes needed?

See above

### Does this PR introduce any user facing changes?

NA

### Memory saves
If there are `n` inodes to process, the save will be:
1. An `ArrayList` of size `n`, so `8 byte * n`
2. One `Pair<>` per inode, so `32 byte * n`
3. Some `ArrayList` resizing internally by initializing with the final
size. So the save is bounded by `8 byte * n`.

The total save is `<(8+32+8) byte * n`. But these are temporary
variables so the cost can be less significant than long-held ones. The
CPU save could be a more significant benefit, subject to our observation
from micro benchmark.

pr-link: Alluxio/alluxio#15052
change-id: cid-783dd408c6d53ad993178473de4914cf4e41f56d