Enable UFS to return Fingerprint
### What changes are proposed in this pull request?

This change adds a method for UFS to directly return a Fingerprint
object, so we don't need to concat a string first then parse it.

### Why are the changes needed?

Save the intermedia cost of String creation and parsing.

### Does this PR introduce any user facing changes?

There's one more method added to UnderFileSystem interface but it has a
default implementation. So NA.

pr-link: Alluxio/alluxio#15180
change-id: cid-b72a5c78f2790003dbab771af1515629fcdab796