Make log compaction local in embedded journal
### What changes are proposed in this pull request?
Making `RaftServerConfigKeys.Log.setPurgeUptoSnapshotIndex` to `true` by
default. This enables each master in the quorum to look only at the
local snapshots it possesses in order to determine which log files to
prune.
Before, masters would only prune based on the quorum consensus of
committed log index.

### Why are the changes needed?
Community members have noticed that Ratis logs grow unbounded in the
presence of a slow master, which can crash certain masters when the UFS
runs out of space in extreme cases.

### Does this PR introduce any user facing changes?
Adds a `PropertyKey` to enable or disable this new feature.

pr-link: Alluxio/alluxio#15245
change-id: cid-7fe8e99f52b170f1bea9027f572b77aae2eea00c