Create MaxFile stressbench
This PR introduces a new stressbench test called MaxFileBench.
MaxFileBench creates 0-byte files until the master becomes unresponsive.
To ensure we are pushing the master to its limits, we pause between a
failure and another attempt at creating a file. This pause doubles in
length on successive failure, and gets reset when a new success occurs.
If the pause gets too large the test ends. This ensures that the master
has actually become unavailable rather than flaky.

pr-link: Alluxio/alluxio#15194
change-id: cid-deb747de72dfd225260cc5a286defc2a20969c2e