Fix some minor benchMark bug and Refactor Master max throughput
fix https://github.com/Alluxio/alluxio/issues/14654

Major changes:
Remove `alluxio.stress.cli.suite.Suite`, which duplicates with the
`Benchmark` class.
Fewer classes make the structure clearer.

The base methods of MasterMaxThroughput and JobServiceMaxThroughput are
merged.

`AbstractMaxThroughputSummary` both interfaces with Summary, TaskResult
to adapt with the `Benchmark`

some bug fix :
1. some unit test may output some interference logs (For example, some
warn logs indicating that the list file is too long).So I added an
identifier to the benchmark output(otherwise, json deserialization to
get the result of each unit test will make an error).

2. The `CREATE_FILES` function in `StressJobServiceBench` cannot be used
(the number of successes is not cumulative, and the delete file step
also fails). The warmup function is also missing

3. When the file system is the HDFS file system, the parent dir can be
deleted to delete the children files. It takes too long to delete the
child files one by one (but this is necessary on s3).

pr-link: Alluxio/alluxio#14819
change-id: cid-d7b1ed6ad66c26f94273c1112797fa1454c55e26