Close JournalServiceClient asynchronously to account for async calls
Currently, the `JournalServiceClient` used in
`SnapshotReplicationManager` is closed using try-resource blocks. This
causes the async calls within said blocks to fail sometimes as the
client is closed when the call is launched.
This PR ensures that the clients are closed asynchronously after the
call is finished using `whenComplete` blocks.
A test accompanies this to ensure proper snapshot propagation.

pr-link: Alluxio/alluxio#15284
change-id: cid-3be650686f0bdca64a1f6205a5fd28b1ef9c7896