Add missing options in DistributedLoad command
### What changes are proposed in this pull request?

Add the missing options to distributed load cli.


### Why are the changes needed?

Options like `--excluded-hosts` etc., are missing in `alluxio fs
distributedLoad` command. The functionalities controlled by the options
are introduced in #13506, but not accessible from the command line
interface. Attempting to run the command with this option results in an
unrecognized option error:

```console
$ bin/alluxio fs distributedLoad --excluded-hosts host1,host2 /tmp
Failed to parse args for distributedLoad: Unrecognized option:
--excluded-hosts
Usage: distributedLoad [--replication <num>] [--active-jobs <num>]
[--batch-size <num>] [--index] [--hosts <host1>,<host2>,...,<hostN>]
[--host-file <hostFilePath>] [--excluded-hosts
<host1>,<host2>,...,<hostN>] [--excluded-host-file <hostFilePath>]
[--locality <locality1>,<locality2>,...,<localityN>] [--locality-file
<localityFilePath>] [--excluded-locality
<locality1>,<locality2>,...,<localityN>] [--excluded-locality-file
<localityFilePath>] [--passive-cache] <path>
Loads a file or all files in a directory into Alluxio space.
```

Note that the option is present in the usage string.

### Does this PR introduce any user facing changes?

Yes, now users can run the command with the new options introduced in
#13506.

pr-link: Alluxio/alluxio#15280
change-id: cid-425962a53e82616609aef0c2751bd33f52f10c1c