Fix alluxio tarball owner in docker
The owner of files in `opt/alluxio` is the owner of the tarball. If the
user builds the tarball locally as non-alluxio user, the user alluxio
(uid 1000) inside the docker container may not have the permission to
write into `alluxio-env.sh`, which is a required step in
`entrypoint.sh`, and thus the cluster is not able to start.

solves #15309

pr-link: Alluxio/alluxio#15320
change-id: cid-29500ac13d48b564150849a65612119747069952