Fix absent cache stats are always zero
### What changes are proposed in this pull request?

Fix the absent cache stats which are always zero.

### Why are the changes needed?

From the doc on `Cache::stats()`:

> Returns a current **snapshot** of this cache's cumulative statistics,
or a set of default values if the cache is not recording statistics. All
statistics begin at zero and never decrease over the lifetime of the
cache.
**Warning**: this cache may not be recording statistical data. For
example, a cache created using CacheBuilder only does so if the
CacheBuilder.recordStats method was called. If statistics are not being
recorded, a CacheStats instance with zero for all values is returned.

This fix

1. enables stats collection on the cache
2. retrieves a new snapshot every time the gauge is polled, instead of
the stale value captured by the lambda.

### Does this PR introduce any user facing changes?

Yes, the metrics are fixed.

pr-link: Alluxio/alluxio#15356
change-id: cid-dcfbb9a2805fdd86f1676d3d227699d3994ccb3e