[DOCFIX] Update quickstart doc
update the list of integrations in the quick start guide. i am proposing
to trim these lists to the ones that Alluxio has a stable integration
with as opposed to an integration that is not thoroughly explored. also
note that if the list is supposed to be thorough and comprehensive, then
it will become quickly out of date as more integrations are added and
these seem to be less important.

pr-link: Alluxio/alluxio#15383
change-id: cid-7965fa8550468eaa9ae6e9d8b635d5b9bb355744