CleanUp createUfsBlockReader
### What changes are proposed in this pull request?

Remove mBlockIdToSessionIds since USER_UFS_BLOCK_READ_CONCURRENCY_MAX is
broken.
Merge openUFSBlock into createUfsBlockReader.
Remove redundant test cases since they are basically duplicates of
UnderFileSystemBlockStoreTest
### Why are the changes needed?

Previously, mBlockIdToSessionIds is introduced to control
UfsReadConcurrency. But basically, this functionality is broken because
when we `createUfsBlockReader` we just drop the value returned from
`openUfsBlock`(line 511 in `DefaultBlockWorker`) and create BlockReader
anyway.
And IMO USER_UFS_BLOCK_READ_CONCURRENCY_MAX is not a good way to control
the concurrency to UFS(the default value is Integer.MAX_VALUE which is
almost restricting nothing).
So we propose to delete this broken functionality instead of picking up
the dropped value. later on, we need to reimplement this throttling to
UFS in a proper way.

Merge openUFSBlock into createUfsBlockReader since openUFSBlock is
always used within createUfsBlockReader and openUFSBlock is just a
combination of UFS fallback logic and acquire access(the name is
confusing as well since it does not actually open block).


### Does this PR introduce any user-facing changes?
na

pr-link: Alluxio/alluxio#15369
change-id: cid-47d9374fc40d27fd97739d5a97ab7a0b8ab0fa00