Improve BlockWorker and LocalBlockStore APIs
1) APIs like accessBlock lockBlock, unlockBlock in BlockWorker are just
wrappers on the corresponding LocalBlockStore APIs that are specific to
LocalBlockStore. Using the LocalBlockStore APIs directly and removing the
wrapper APIs in BlockWorker.

2) Rename LocalBlockStore.lockBlock to pinBlock. The API is supposed to
be indicating future block read access and is intended to avoid eviction.
Currently the underlying implementation in TieredBlockStore is still using
locking and the API still returns lockId. But changing the API to indicate
that such strong guarantee is not necessary for all implementations. In the
future this function should just return void.

3) Rename LocalBlockStore.unlockBlock to unpinBlock and remove Exceptions not thrown.

4) Remove BlockWorker.INVALID_LOCK_ID. Using OptionalLong to represent lockId.
Invalid lock is represented as OptionalLong.empty(), making special case
handling explicit.