Remove getBlockMeta from BlockWorker and LocalBlockStore
### What changes are proposed in this pull request?

Remove getBlockMeta from BlockWorker and LocalBlockStore.

### Why are the changes needed?

There are getBlockMeta and getVolatileBlockMeta. The difference is
getBlockMeta requires a sessionId and blockId, and validates the lock.
This is more like a debugging feature. Since the caller has already
acquired
the lock, calling getVolatileBlockMeta is always safe in this situation,
and
avoids the overhead of validating the lock.

pr-link: Alluxio/alluxio#15447
change-id: cid-d988fcf91ac888b0c950a962d9fc80eee5008133