Consolidate static analysis checks from microbench
When microbench was introduced in #15518, it required new suppressions
for the checkstyle plugin and the spotbugs plugin. This PR consolidates
the new suppressions in the existing top level files.
This makes suppressions more understandable as they are all found in one
single file, and also helps IDE extensions function properly (as there
is only one file they must look at).

pr-link: Alluxio/alluxio#15572
change-id: cid-49f507d007afefcb67534d6ea03a8f3165d58fe9