Improve accuracy of the metrics on num of pages
### What changes are proposed in this pull request?
As Baolong and I discussed in
https://github.com/Alluxio/alluxio/pull/14585, I propose the changes
below to improve the accuracy of the metrics.
Using the actual map size instead of the inc/dec for the metrics of num
of pages
Remove the unnecessary field mPages

### Why are the changes needed?

Please clarify why the changes are needed. For instance,
This change would improve the accuracy of the metrics. Because the
exceptions might skip some of the inc/dec which might make the value of
the metrics diverged greatly.

### Does this PR introduce any user facing changes?

None

pr-link: Alluxio/alluxio#15121
change-id: cid-fb0463c839a728b15d4d12ca9b20a84e011c857a