Fix wrong abs method
In the current implementation, the value of the `sleepMs` variable is
almost impossible to be `Integer.MIN_VALUE`, because the value of
`mBaseSleepTimeMs` is 50 or 100, neither is a power of 2.

As can be seen from the name of this method (`"abs"`), its function
is mainly to prevent negative numbers from appearing during overflow,
so it is OK to use `Math.abs()` directly.