Optimize the checks of the loop in getMountPoint
### What changes are proposed in this pull request?

According to the implementation of `getMountPoint`, it receives an
alluxio uri as the parameter, gets the `path` from this uri by calling
`uri.getPath()`, iterates the whole mount table's key-value pairs to
find the key that is the nearest parent directory of `path`, which in
other words is the longest prefix of `path`.

To make sure that the longest prefix can be recorded in `lastMount`, the
condition is as follow:
```java
// ...
for (Map.Entry<String, MountInfo> entry :
mState.getMountTable().entrySet()) {
String mount = entry.getKey();
if (!mount.equals(ROOT) && PathUtils.hasPrefix(path, mount)
&& PathUtils.hasPrefix(mount, lastMount)) {
lastMount = mount;
}
}
return lastMount;
```
`mount` is the current looped entry's key. The first two && condition
check if the `mount` is the prefix of `path`, and the third condition
checks if `lastMount` is the prefix of `mount`. However, since when
executing the third condition, `mount` is already the prefix of `path`,
so if the `mount` is longer than `lastMount`, `mount` is sure to be the
path that is more close to `path` than `lastMount`. Therefore, it is no
need to call `hasPrefix` and we can instead simply compares the length
of `mount` and `lastMount`.

I also removed the redundant `try-catch` block in `hasPrefix`
### Why are the changes needed?

To reduce unnecessary function call that will be called many times in a
loop, in order to speed up `getMountPoint`.

### Does this PR introduce any user facing changes?
No

pr-link: Alluxio/alluxio#15300
change-id: cid-d882405bc2f79dd1c4ed73228f50e271fb4905ac