Remove Java doc for private functions
### What changes are proposed in this pull request?

change the Java doc for private functions into comment if they are still
useful. removed if information is trivil.

### Why are the changes needed?

We shouldn't write javadoc for private methods. End users don't have
access to private fields or methods so there really isn't a point in
providing javadoc for them. Private fields and methods are only meant
for the developer. So leaving them as comments if seem useful.

### Does this PR introduce any user facing changes?

No

pr-link: Alluxio/alluxio#15627
change-id: cid-15c173001c21e7df14d7e71b9f2b56cc87f03bc8