Fix max reservoir metric initialization
### What changes are proposed in this pull request?

The max reservoir metric is now initialized to 0.

### Why are the changes needed?

Previously the max reservoir metric was initialized to Long.MIN_VALUE
which is a negative value, but the max reservoir is used to track the
max size of something that should be initialized to 0.

### Does this PR introduce any user facing changes?

Any metric that uses the max reservoir (currently the thread pool
executor monitoring the metadata queues) will now be initialized to 0.

pr-link: Alluxio/alluxio#15600
change-id: cid-3f9dbf883bb595a7731b034c7d95ce6e9877b1d5