Fix wrong abs method in ExponentialBackoffRetry
### What changes are proposed in this pull request?
Fix wrong abs method in `ExponentialBackoffRetry`.

### Why are the changes needed?

In the current implementation, the value of the `sleepMs` variable is
almost impossible to be `Integer.MIN_VALUE`, because the value of
`mBaseSleepTimeMs` is 50 or 100, neither is a power of 2.

As can be seen from the name of this method (`"abs"`), its function
is mainly to prevent negative numbers from appearing during overflow,
so it is OK to use `Math.abs()` directly.

### Does this PR introduce any user facing changes?
No

pr-link: Alluxio/alluxio#15585
change-id: cid-735e9ceb6ac5e02d0977d0da3b8a3bae2c08c86f