Remove System.exit() calls in ufs journal catchup and propagate errors
### What changes are proposed in this pull request?
The `UfsJournalCheckpointThread` calls `System.exit(-1)` when
encountering journal corruption. This causes a deadlock between the
`UfsJournalCheckpointThread` and the shutdown hook calling
`AlluxioMasterProcess.stop()`.
This PR removes the `System.exit()` calls in favor of propagating
exceptions when the thread completes.
The net effect is that the master process completely releases its
resources when shutting down and terminates without having to be issued
`kill -9`.

### Why are the changes needed?
For clean master process termination on errors.

### Does this PR introduce any user facing changes?
No.

pr-link: Alluxio/alluxio#15699
change-id: cid-055a9c1a668761332c61717c59174a86e2cbce17