Rename BlockStore interface to BlockMetaStore
### What changes are proposed in this pull request?

Rename BlockStore interface to BlockMetaStore

### Why are the changes needed?
The blockStore interface actually is a store for the metadata of blocks
rather than the data of blocks, so New name may reduce confusion.

Also I would like to use the name of BlockStore for the abstraction of
pagedBlockStore and tieredBlockStore.

### Does this PR introduce any user facing changes?
N/A

pr-link: Alluxio/alluxio#15693
change-id: cid-2b0de2733e9b3fad84ce017c81dc0e75ade84187