Simplify Grpc Channel implementation
Previously there GrpcChannelKey and GrpcConnectionKey, also
GrpcChannel and GrpcConnection. The distinction between them are
not clear. It seems that GrpcConnectionKey is GrpcChannelKey + groupIndex
(connection key has a channel key). GrpcChannel is GrpcConnection +
authentication (channel has a connection). This logic is very confusing.
Simplify the logic by merging GrpcChannelKey and GrpcConnectionKey into
GrpcChannelKey, which is the equivalent of the previous GrpcConnectionKey.
Also merging GrpcChannel and GrpcConnection into GrpcChannel, which is
equivalent to previous GrpcChannel. Also renaming GrpcConnectionPool to
GrpcChannelPool. Removing grpc connection to reduce confusion.