Refactor fuse options
Major changes summary:
1. Create `AlluxioFuseCliOpts` to parse and store the command line
information.
2. Replace `FuseMountConfig` with`AlluxioFuseFileSystemOpts`. It checks
`AlluxioFuseCliOpts` and `AlluxioConfiguration` and do some formatting
to get the final options for Alluxio fuse.
3. Delete propertyKey `FUSE_MAXWRITE_BYTES`. First, if user doesn't set
it, the default value we set is the same the default set by Libfuse.
Second, if users want to set it, they should do it in propertykey
`FUSE_MOUNT_OPTIONS` or through command line.

pr-link: Alluxio/alluxio#15553
change-id: cid-593cc463e5c0687502e17745956e8057b886aaa4