Optimize the access speed of the inAlluxioData UI page
### What changes are proposed in this pull request?

To optimize the access speed of the inAlluxioData UI page

### Why are the changes needed?

When there are many files in alluxio, if we open the inAlluxio web ui of
the Alluxio master, it will take a long time and cause the Alluxio
master to freeze. We may specify a certain number of file information to
be displayed on the inAlluxio web through configuration items to avoid
the above situation.

### Does this PR introduce any user facing changes?
No

pr-link: Alluxio/alluxio#15716
change-id: cid-103009cad3708a80b7d0080564a2a051c47f5c51