Add liveness probe to fuse pod
If a lot of fuse pods are spawned simultaneously, some mount points
inside the pods will be corrupted that user cannot do `ls` or `cd` to
it. No error is reported by either Alluxio or libfuse. A simple remount
is able to solve the problem.

The liveness probe will check whether the mount point is corrupted, and
if so, restart the container.

This should be able to solve
https://github.com/Alluxio/alluxio/issues/15679.

pr-link: Alluxio/alluxio#15703
change-id: cid-d580eb031ef718a6ed21c3e24f0d425a301be6f9