Fix wrong pool when releasing BWC
### What changes are proposed in this pull request?

Fix `IllegalArgumentException` when releasing a block worker client
because it was released to the wrong pool.

### Why are the changes needed?

Fix #15556.

On the creation of the block worker and its release, the key used to
retrieve the pool from the map are computed separately, and its
ingredients could have changed in between. Therefore, when releasing the
client, the wrong pool may be retrieved. This can be due to a number of
reasons, for example the Alluxio configuration has changed during this
time, and the client picks up new conf via client master config sync.

The fix is to embed the key inside the closable resource to avoid
computing it a second time.

### Does this PR introduce any user facing changes?

No.

pr-link: Alluxio/alluxio#15579
change-id: cid-c593348a3ddaf29f702289ee617737dd4de15bba