Install fuse3 in jenkins image for github action
The current JNIFuseIntegrationTest is always running fuse2 even it
claims it runs with fuse3. The PR
https://github.com/Alluxio/alluxio/pull/15863 fixes this by replacing
the existing incorrect loading libfuse process for the test with a
correct one, but the jenkins image doesn't have fuse3, so the test
hangs.

This PR is to install Fuse3 on the jenkins image. On Ubuntu used by the
jenkins image, package `fuse3` installs `fuse2` as well, which is why
only `fuse3` is installed.

pr-link: Alluxio/alluxio#15876
change-id: cid-1d29745c40ec02b8c16263079c4df91196565f51