Allow Docker image file permissions to be modified at runtime
### What changes are proposed in this pull request?

Rrefactor the Dockerfile's entrypoint script so that we can set
permissions on a provided env/argument instead.

### Why are the changes needed?

Please clarify why the changes are needed. For instance,

Currently in the Alluxio Dockerfile the USER, UID and GID are all
hard-coded. Thus only way to change the file permissions is to recompile
a Docker image with new values for the arguments.

This creates a problem for users of the Helm chart who try to modify the
user / UID used to run the Docker container in k8s. The corresponding
file permissions do not get updated to the new UID.

I suggest we should consider refactoring the Dockerfile's entrypoint
script so that we can set permissions on a provided env/argument
instead.

### Does this PR introduce any user facing changes?

No, it doesn't.

pr-link: Alluxio/alluxio#15898
change-id: cid-8f27342bc5e3d59ae33e5be35321eede47bbf3d0