Assign version to BlockWorkerClientService
### What changes are proposed in this pull request?
- Reassign `BlockWorkerClientServiceHandler` of type
`ServiceType#BLOCK_WORKER_CLIENT_SERVICE` and deprecate
`ServiceType#FILE_SYSTEM_WORKER_WORKER_SERVICE`
- Assign a version to `BLOCK_WORKER_CLIENT_SERVICE`

### Why are the changes needed?
- `FILE_SYSTEM_WORKER_WORKER_SERVICE` was a bad name.
- `BlockWorkerClient` did not check service version, because it did not
extend `AbstractClient` which does the version check. After #15837,
`BlockWorkerClient` can be unified to the check version mechanism of
`AbstractClient`.

### Does this PR introduce any user facing changes?

#### Impact on Client Backwards Compatibility

This particular PR and its downstream
https://github.com/Alluxio/alluxio/pull/15837 will have the following
impact on backwards compatibility:

##### Before

- An old worker has a version service under the name
FILE_SYSTEM_WORKER_WORKER_SERVICE and an unknown version number (-1).
- An old BlockWorkerClient does not check the worker's version after
connecting to the worker.

##### After

- A new worker has a version service under the name
BLOCK_WORKER_CLIENT_SERVICE and a dedicated version number (1).
- A new client will check the worker's version and throw an exception if
the versions are not equal.

Therefore, the breaking change is when a new client connects to an old
worker, an unexpected exception will make the client refuse to connect
to the worker.

pr-link: Alluxio/alluxio#15874
change-id: cid-64dd3ad46a58f51a68af9a206f057237926dc261