Fix ActiveSync timeout retry
### What changes are proposed in this pull request?

https://github.com/Alluxio/alluxio/pull/15765 changed the ActiveSync
retry mechanism to use reuse a single RetryPolicy object for every call.
This caused the ActiveSync calls to not work correctly as the timer
object in this single object was always expired after every call. This
reverts the change, but keeps the cleanup restructuring from that PR.

### Why are the changes needed?

This fixes ActiveSync calls timing out unnecessarily.

### Does this PR introduce any user facing changes?

No

pr-link: Alluxio/alluxio#15891
change-id: cid-3c19803134e38fed4b04581fcf66702f88a78b54